refactor : Entity의 id 컬럼명 지정 및 NoArgsConstructor 접근 제어자 변경
