+++ batch/src/main/java/org/meotppo/webti/job/processor/ProcessorConfig.java
@@ -2,8 +2,8 @@ package org.meotppo.webti.job.processor;
 
 import lombok.RequiredArgsConstructor;
 import org.meotppo.webti.domain.entity.jpa.profile.Profile;
-import org.meotppo.webti.domain.entity.jpa.statistic.Statistic;
-import org.meotppo.webti.domain.entity.mongo.testresult.TestResult;
+import org.meotppo.webti.domain.entity.jpa.result.Statistic;
+import org.meotppo.webti.domain.entity.mongo.result.TestResult;
 import org.meotppo.webti.domain.repository.jpa.developertype.ProfileRepository;
 import org.springframework.batch.item.ItemProcessor;
 import org.springframework.context.annotation.Bean;
@@ -24,7 +24,7 @@ public class ProcessorConfig {
                             () -> new IllegalArgumentException("No profile found for type: " + item.getMbtiType()));
 
             return Statistic.builder()
-                    .developerProfile(profile)
+                    .profile(profile)
                     .count(1L)
                     .matchCount(item.isMatch() ? 1L : 0L)
                     .build();
