+++ batch/src/main/java/org/meotppo/webti/job/reader/ReaderConfig.java
@@ -1,7 +1,7 @@
 package org.meotppo.webti.job.reader;
 
 import java.util.Collections;
-import org.meotppo.webti.domain.entity.mongo.testresult.TestResult;
+import org.meotppo.webti.domain.entity.mongo.result.TestResult;
 import org.springframework.batch.item.data.MongoPagingItemReader;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
