+++ batch/src/main/java/org/meotppo/webti/job/writer/WriterConfig.java
@@ -1,6 +1,6 @@
 package org.meotppo.webti.job.writer;
 
-import org.meotppo.webti.domain.entity.jpa.statistic.Statistic;
+import org.meotppo.webti.domain.entity.jpa.result.Statistic;
 import org.meotppo.webti.domain.repository.jpa.statistic.StatisticRepository;
 import org.springframework.batch.item.ItemWriter;
 import org.springframework.context.annotation.Bean;
@@ -14,9 +14,9 @@ public class WriterConfig {
     @Bean(name = STATISTIC_WRITER)
     public ItemWriter<Statistic> statisticWriter(StatisticRepository statisticRepository) {
         return items -> items.forEach(item -> {
-            Statistic existing = statisticRepository.findByDeveloperProfile(item.getDeveloperProfile())
+            Statistic existing = statisticRepository.findByProfile(item.getProfile())
                     .orElseThrow(() -> new RuntimeException(
-                            "Statistic not found for developer profile: " + item.getDeveloperProfile()));
+                            "Statistic not found for developer profile: " + item.getProfile()));
             existing.updateCount(existing.getCount() + item.getCount());
             existing.updateMatchCount(existing.getMatchCount() + item.getMatchCount());
         });
