+++ domain/src/main/java/org/meotppo/webti/domain/entity/jpa/question/Question.java
@@ -8,6 +8,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
 import java.util.List;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,10 +16,11 @@ import org.meotppo.webti.domain.entity.jpa.common.JpaEntityDate;
 
 @Getter
 @Entity
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Question extends JpaEntityDate { // 질문과 답변을 프론트로 전달 -> 프론트는 질문에 대한 답변 선택, 점수 계산 후 백엔드로 전송
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "question_id")
     private Long id;
 
     @Column(length = 500, nullable = false)
