+++ domain/src/main/java/org/meotppo/webti/domain/repository/mongo/testresult/TestResultRepository.java
@@ -1,6 +1,6 @@
 package org.meotppo.webti.domain.repository.mongo.testresult;
 
-import org.meotppo.webti.domain.entity.mongo.testresult.TestResult;
+import org.meotppo.webti.domain.entity.mongo.result.TestResult;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface TestResultRepository extends MongoRepository<TestResult, String> {
